+++ src/main/java/pa/model/CountryRepository.java
@@ -12,7 +12,7 @@ public interface CountryRepository extends CrudRepository<Country, String> {
 
 	List<Country> findByName(String name);
 	List<Country> findByRegion(Region region);
-	List<Country> findByRegion_id(int id);
+	List<Country> findByRegion_id(long id);
 
 	List<Country> findByNameLikeIgnoreCase(String name);
 }
